requirements.txt
fuse-python>=1.0.0
web3>=6.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
